# api/openapi.yaml
openapi: 3.0.3
info:
  title: HallOfFame API
  description: 名人堂 - QQ 群言论展示系统 API 文档
  version: 1.0.0
servers:
  - url: http://localhost:8888/api/v1
    description: 主服务器
  - url: http://localhost:8889/api/v1/bot
    description: Bot 服务器 (无鉴权)
paths:
...
  /admin/users:
    get:
      summary: 用户列表
      tags: [管理员]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 用户列表
  /admin/users/{id}/role:
    put:
      summary: 修改用户角色
      tags: [管理员]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 角色更新成功
  /admin/whitelist:
    get:
      summary: 白名单列表
      tags: [管理员]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 白名单列表

    post:
      summary: 添加白名单
      tags: [管理员]
      security:
        - BearerAuth: []
      responses:
        '201':
          description: 添加成功

    delete:
      summary: 移除白名单
      tags: [管理员]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 移除成功
  /admin/login-logs:
    get:
      summary: 登录日志
      tags: [管理员]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 登录日志列表

  # ============ Bot (无鉴权) ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT